fix numpy integer requirement for index

+++ peakdetect.py
@@ -13,7 +13,7 @@ def nzc_temporal_filtered(fs, waveform, min_spacing=0.3):
 
 def nzc_noise_filtered(fs, waveform, dev=1.0, min_spacing=0.3):
 
-    min_noise = waveform[:1e-3*fs].std()*dev
+    min_noise = waveform[:int(1e-3*fs)].std()*dev
 
     p_ind = nzc(waveform)
     n_ind = nzc(-waveform)
